helpdeskproject/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
crispy-bootstrap5==0.6
cssselect2==0.4.1
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-import-export==2.7.1
et-xmlfile==1.1.0
fonttools==4.28.1
html5lib==1.1
idna==3.3
iniconfig==1.1.1
kiwisolver==1.3.2
MarkupPy==1.14
matplotlib==3.4.3
mysqlclient==2.0.3
numpy==1.21.4
odfpy==1.4.1
opencv-python==4.5.4.58
openpyxl==3.0.9
packaging==21.3
parinya==0.0.27
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.4
pyphen==0.11.0
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.7
weasyprint==53.3
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.1.9
